Add files via upload
Here is a presentation explaining the segmentation protocol using SuRVoS and imageJ.
The imageJ macro in imageJ macro language (.ijm) is in the other document. It can be copied into the imageJ macro edit window and contains instructions for applying the macro to different data in which blob-shaped objects can be segmented from the background, counted and measured.